==> notebooks/7.0-doccano_export.ipynb <==
## Doccano Processing

Processes the Doccano annotated dataset and format it for Excel for final quality control and further coding refinement.

import os
import json
import pandas as pd

# load JSON:
doccano_file = "final_annotated_idw.json"

with open(
    os.path.join(
        "..",
        "doccano",
        "exports",
        doccano_file
    ),
    "r"
) as f:
    data = json.load(f)
 
 # pass to Pandas:   
df = pd.DataFrame(data)

# join the 'Comments' values:
df["Comments"] = df["Comments"].apply(lambda row: " ".join(row))

# rename `id` for clarity:
df.rename(columns={"id": "doccano_id"}, inplace=True)

df.head()

**Get dummy values for labels:**

dummies = pd.get_dummies(df.explode("label")["label"]).groupby(level=0).sum()
df = pd.concat([df, dummies], axis=1)
df["label"] = df["label"].apply(lambda row: "; ".join(row))

df.head()

**Pass to Excel for final coding refinement and quality control:**

outfile = "coded_sample.xlsx"
df.to_excel(
    os.path.join(
        "..",
        "data",
        "coding",
        "raw",
        outfile
    ),
    index=False,
    sheet_name="doccano_sample"
)

`<-- END -->`
